Tema04 - Flask Autenticación/Clase09/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
psycopg2-binary
python-dotenv
cryptography
flask-mail